requirements.txt
crewai==0.11.0
crewai-tools==0.12.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
python-dotenv==1.0.0
